Updated to calculate all at the beginning
I decided to abandon this method as it was using too much memory, crashing RStudio, and taking forever to run, likely due to data frame nesting and the for loop within the while loop. I.e. This method is ridiculously inefficient compared to the calculate-as-you-go method. The code that's been uploaded has minor changes in order to run faster and use less memory. Currently, it draws the lines, but incorrectly. View at your own risk.